refactor: global stats components
